requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.3.0
python-decouple==3.8
requests==2.32.5